Move CSS next to component
