+++ components/testimonials/TestimonialRow.css
@@ -1,0 +1,11 @@
+.testimonialBubble:after {
+	content: '';
+	position: absolute;
+	bottom: -16px;
+	left: 15px;
+	width: 0;
+	height: 0;
+	border: 20px solid transparent;
+	border-top-color: rgb(242 242 242);
+	border-bottom: 0;
+}
