+++ styles.css
@@ -16,14 +16,3 @@
   @import "react-responsive-carousel/lib/styles/carousel.min.css";
 }
 
-.testimonialBubble:after {
-	content: '';
-	position: absolute;
-	bottom: -16px;
-	left: 15px;
-	width: 0;
-	height: 0;
-	border: 20px solid transparent;
-	border-top-color: rgb(242 242 242);
-	border-bottom: 0;
-}
